Put PID and feedforward logic into `Pivot.java`
Additionally, change to set desired angle instead of setting speed.

+++ src/main/java/frc/robot/Constants.java
@@ -137,8 +137,8 @@ public final class Constants {
     public static final double KV = 0;
     public static final double KA = 0;
 
-    public static final int MAX_VELOCITY = 0;
-    public static final int MAX_ACCELERATION = 0;
+    public static final double MAX_VELOCITY = 0;
+    public static final double MAX_ACCELERATION = 0;
     public static final int MAX_CURRENT = 40;
   }
 }
